+++ contracts/HoldefiSettings.sol
@@ -19,45 +19,45 @@ contract HoldefiSettings is HoldefiOwnable {
 
 	using SafeMath for uint256;
 
-	uint constant public rateDecimals = 10 ** 4;
+	uint256 constant public rateDecimals = 10 ** 4;
 
-	uint constant public periodBetweenUpdates = 864000;
+	uint256 constant public periodBetweenUpdates = 864000;
 
-	uint constant public maxBorrowRate = 4000;				  //40%
+	uint256 constant public maxBorrowRate = 4000;				  //40%
 
-	uint constant public borrowRateMaxIncrease = 500;		  //5%
+	uint256 constant public borrowRateMaxIncrease = 500;		  //5%
 
-	uint constant public minSuppliersShareRate = 5000;		  //50%
+	uint256 constant public minSuppliersShareRate = 5000;		  //50%
 
-	uint constant public suppliersShareRateMaxDecrease = 500; //5%
+	uint256 constant public suppliersShareRateMaxDecrease = 500; //5%
 
-	uint constant public maxValueToLoanRate = 20000; 		  //200%
+	uint256 constant public maxValueToLoanRate = 20000; 		  //200%
 
-	uint constant public valueToLoanRateMaxIncrease = 500;	  //5%
+	uint256 constant public valueToLoanRateMaxIncrease = 500;	  //5%
 
-	uint constant public maxPenaltyRate = 13000; 			  //130%
+	uint256 constant public maxPenaltyRate = 13000; 			  //130%
 
-	uint constant public penaltyRateMaxIncrease = 500; 		  //5%
+	uint256 constant public penaltyRateMaxIncrease = 500; 		  //5%
 
 	// Markets Features 
 	struct MarketSettings {
 		bool isActive;
 
-		uint borrowRate;
-		uint borrowRateUpdateTime;
+		uint256 borrowRate;
+		uint256 borrowRateUpdateTime;
 
-		uint suppliersShareRate;
-		uint suppliersShareRateUpdateTime;	
+		uint256 suppliersShareRate;
+		uint256 suppliersShareRateUpdateTime;	
 	}
 
 	// Collaterals Features
 	struct CollateralSettings {
 		bool isActive;
-		uint valueToLoanRate;   // Collateral liquidation threshold
-		uint VTLUpdateTime;
-		uint penaltyRate; 		// Portion of collateral being liquidated during liquidation
-		uint penaltyUpdateTime;
-		uint bonusRate;		    // Bonus for buyers who buy liquidated collaterals
+		uint256 valueToLoanRate;   // Collateral liquidation threshold
+		uint256 VTLUpdateTime;
+		uint256 penaltyRate; 		// Portion of collateral being liquidated during liquidation
+		uint256 penaltyUpdateTime;
+		uint256 bonusRate;		    // Bonus for buyers who buy liquidated collaterals
 	}
 
 	// Asset address => Market features 
@@ -69,23 +69,23 @@ contract HoldefiSettings is HoldefiOwnable {
 
 	HoldefiInterface public holdefiContract;
 
-	event BorrowRateChanged(address market, uint newRate);
+	event BorrowRateChanged(address market, uint256 newRate);
 
-	event SuppliersShareRateChanged(address market, uint newRate);
+	event SuppliersShareRateChanged(address market, uint256 newRate);
 
 	event MarketAdded(address market);
 
 	event MarketRemoved(address market);
 
-	event CollateralAdded(address collateral, uint valueToLoanRate, uint penaltyRate, uint bonusRate);
+	event CollateralAdded(address collateral, uint256 valueToLoanRate, uint256 penaltyRate, uint256 bonusRate);
 
 	event CollateralRemoved(address collateral);
 
-	event ValueToLoanRateChanged(address collateral, uint newRate);
+	event ValueToLoanRateChanged(address collateral, uint256 newRate);
 
-	event PenaltyRateChanged(address collateral, uint newRate);
+	event PenaltyRateChanged(address collateral, uint256 newRate);
 
-	event BonusRateChanged(address collateral, uint newRate);
+	event BonusRateChanged(address collateral, uint256 newRate);
 	
 	constructor (address ownerChanger) HoldefiOwnable(ownerChanger) public {
 	}
@@ -98,15 +98,15 @@ contract HoldefiSettings is HoldefiOwnable {
 
 	// Returns supply rate and borrow rate 
 	// supply rate = ((total borrow * Borrow rate * suppliers share rate) / total supply
-	function getInterests (address market, uint totalSupply, uint totalBorrow) external view returns(uint borrowRate, uint supplyRate) {
+	function getInterests (address market, uint256 totalSupply, uint256 totalBorrow) external view returns(uint256 borrowRate, uint256 supplyRate) {
 		borrowRate = marketAssets[market].borrowRate;
-		uint suppliersShareRate = marketAssets[market].suppliersShareRate;
+		uint256 suppliersShareRate = marketAssets[market].suppliersShareRate;
 		if (totalSupply == 0){
 			supplyRate = 0;
 		}
 		else {
-			uint totalInterestFromBorrow = totalBorrow.mul(borrowRate);
-			uint suppliersShare = totalInterestFromBorrow.mul(suppliersShareRate);
+			uint256 totalInterestFromBorrow = totalBorrow.mul(borrowRate);
+			uint256 suppliersShare = totalInterestFromBorrow.mul(suppliersShareRate);
 			suppliersShare = suppliersShare.div(rateDecimals);
 			supplyRate = suppliersShare.div(totalSupply);
 		}
@@ -133,14 +133,14 @@ contract HoldefiSettings is HoldefiOwnable {
 	}
 	
 	// Owner can set a new borrow rate
-	function setBorrowRate (address market, uint newBorrowRate) external onlyOwner {
+	function setBorrowRate (address market, uint256 newBorrowRate) external onlyOwner {
 		require (newBorrowRate <= maxBorrowRate,'Rate should be less than max');
-		uint currentTime = block.timestamp;
+		uint256 currentTime = block.timestamp;
 
 		if (newBorrowRate > marketAssets[market].borrowRate){
-			uint deltaTime = currentTime.sub(marketAssets[market].borrowRateUpdateTime);
+			uint256 deltaTime = currentTime.sub(marketAssets[market].borrowRateUpdateTime);
 			require (deltaTime >= periodBetweenUpdates,'Increasing rate is not allowed at this time');
-			uint maxIncrease = marketAssets[market].borrowRate.add(borrowRateMaxIncrease);
+			uint256 maxIncrease = marketAssets[market].borrowRate.add(borrowRateMaxIncrease);
 			require (newBorrowRate <= maxIncrease,'Rate should be increased less than max allowed');
 		}
 
@@ -154,14 +154,14 @@ contract HoldefiSettings is HoldefiOwnable {
 	}
 
 	// Owner can set a new 'suppliers share rate' (Supplier's share of borrower's interest).
-	function setSuppliersShareRate (address market, uint newSuppliersShareRate) external onlyOwner {
+	function setSuppliersShareRate (address market, uint256 newSuppliersShareRate) external onlyOwner {
 		require (newSuppliersShareRate >= minSuppliersShareRate && newSuppliersShareRate <= rateDecimals,'Rate should be in allowed range');
-		uint currentTime = block.timestamp;
+		uint256 currentTime = block.timestamp;
 
 		if (newSuppliersShareRate < marketAssets[market].suppliersShareRate) {
-			uint deltaTime = currentTime.sub(marketAssets[market].suppliersShareRateUpdateTime);
+			uint256 deltaTime = currentTime.sub(marketAssets[market].suppliersShareRateUpdateTime);
 			require (deltaTime >= periodBetweenUpdates,'Decreasing rate is not allowed at this time');
-			uint maxDecrease = marketAssets[market].suppliersShareRate.sub(suppliersShareRateMaxDecrease);
+			uint256 maxDecrease = marketAssets[market].suppliersShareRate.sub(suppliersShareRateMaxDecrease);
 			require (newSuppliersShareRate >= maxDecrease,'Rate should be decreased less than max allowed');
 		}
 
@@ -174,7 +174,7 @@ contract HoldefiSettings is HoldefiOwnable {
 	}
 
 	// Owner can add a new asset as a market.
-	function addMarket (address market, uint borrowRate, uint suppliersShareRate) external onlyOwner {
+	function addMarket (address market, uint256 borrowRate, uint256 suppliersShareRate) external onlyOwner {
 		require(!marketAssets[market].isActive, "Market exists");
 		require (borrowRate <= maxBorrowRate
 			&& suppliersShareRate >= minSuppliersShareRate
@@ -188,7 +188,7 @@ contract HoldefiSettings is HoldefiOwnable {
 		marketAssets[market].suppliersShareRateUpdateTime = block.timestamp;
 	
 		bool exist = false;
-		for (uint i=0; i<marketsList.length; i++) {
+		for (uint256 i=0; i<marketsList.length; i++) {
 			if (marketsList[i] == market){
 				exist = true;
 				break;
@@ -209,7 +209,7 @@ contract HoldefiSettings is HoldefiOwnable {
 	}
 
 	// Owner can add a collateral asset with its VTL, penalty and bonus rate
-	function addCollateral (address collateralAsset, uint valueToLoanRate, uint penaltyRate, uint bonusRate) external onlyOwner {
+	function addCollateral (address collateralAsset, uint256 valueToLoanRate, uint256 penaltyRate, uint256 bonusRate) external onlyOwner {
 		require(!collateralAssets[collateralAsset].isActive, "Collateral exists");		
 		require (valueToLoanRate <= maxValueToLoanRate
 				&& penaltyRate <= maxPenaltyRate
@@ -235,16 +235,16 @@ contract HoldefiSettings is HoldefiOwnable {
 	}
 	
 	// Owner can set a new VTL rate (Liquidation threshold) for each collateral asset
-	function setValueToLoanRate (address collateralAsset, uint newValueToLoanRate) external onlyOwner {
+	function setValueToLoanRate (address collateralAsset, uint256 newValueToLoanRate) external onlyOwner {
 		require (newValueToLoanRate <= maxValueToLoanRate
 				&& collateralAssets[collateralAsset].penaltyRate <= newValueToLoanRate
 				,'Rate should be in allowed range');
 		
-		uint currentTime = block.timestamp;
+		uint256 currentTime = block.timestamp;
 		if (newValueToLoanRate > collateralAssets[collateralAsset].valueToLoanRate) {
-			uint deltaTime = currentTime.sub(collateralAssets[collateralAsset].VTLUpdateTime);
+			uint256 deltaTime = currentTime.sub(collateralAssets[collateralAsset].VTLUpdateTime);
 			require (deltaTime >= periodBetweenUpdates,'Increasing rate is not allowed at this time');
-			uint maxIncrease = collateralAssets[collateralAsset].valueToLoanRate.add(valueToLoanRateMaxIncrease);
+			uint256 maxIncrease = collateralAssets[collateralAsset].valueToLoanRate.add(valueToLoanRateMaxIncrease);
 			require (newValueToLoanRate <= maxIncrease,'Rate should be increased less than max allowed');
 		}
 	    collateralAssets[collateralAsset].valueToLoanRate = newValueToLoanRate;
@@ -254,17 +254,17 @@ contract HoldefiSettings is HoldefiOwnable {
 	}
 
 	// Owner can set penalty rate for each collateral asset
-	function setPenaltyRate (address collateralAsset ,uint newPenaltyRate) external onlyOwner {
+	function setPenaltyRate (address collateralAsset ,uint256 newPenaltyRate) external onlyOwner {
 		require (newPenaltyRate <= maxPenaltyRate
 				&& newPenaltyRate <= collateralAssets[collateralAsset].valueToLoanRate
 				&& collateralAssets[collateralAsset].bonusRate <= newPenaltyRate
 				,'Rate should be in allowed range');
 
-		uint currentTime = block.timestamp;
+		uint256 currentTime = block.timestamp;
 		if (newPenaltyRate > collateralAssets[collateralAsset].penaltyRate){
-			uint deltaTime = currentTime.sub(collateralAssets[collateralAsset].penaltyUpdateTime);
+			uint256 deltaTime = currentTime.sub(collateralAssets[collateralAsset].penaltyUpdateTime);
 			require (deltaTime >= periodBetweenUpdates,'Increasing rate is not allowed at this time');
-			uint maxIncrease = collateralAssets[collateralAsset].penaltyRate.add(penaltyRateMaxIncrease);
+			uint256 maxIncrease = collateralAssets[collateralAsset].penaltyRate.add(penaltyRateMaxIncrease);
 			require (newPenaltyRate <= maxIncrease,'Rate should be increased less than max allowed');
 		}
 	    collateralAssets[collateralAsset].penaltyRate  = newPenaltyRate;
@@ -274,7 +274,7 @@ contract HoldefiSettings is HoldefiOwnable {
 	}
 
 	// Owner can set bonus rate for each collateral asset
-	function setBonusRate (address collateralAsset, uint newBonusRate) external onlyOwner {
+	function setBonusRate (address collateralAsset, uint256 newBonusRate) external onlyOwner {
 		require (newBonusRate <= collateralAssets[collateralAsset].penaltyRate
 				&& newBonusRate >= rateDecimals
 				,'Rate should be in allowed range');
